% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.as.R
\name{pattern.as}
\alias{pattern.as}
\title{Pattern : Responses in ascending order}
\usage{
pattern.as(data, cols)
}
\arguments{
\item{data}{The data set with the scored responses in the same order as it is administered}

\item{cols}{The columns belonging to the psychometric scale/scales under scrutiny}
}
\value{
The participants (rows) following the pattern, and
the list of said participants along with their responses for all the items in the scale
}
\description{
Pattern : Responses in ascending order
}
\examples{
data <- data.frame(
    Q1 = c(1, 2, 1, 2),
    Q2 = c(2, 1, 2, 1),
    Q3 = c(1, 2, 1, 2),
    Q4 = c(2, 1, 2, 1)
)
pattern.as(data, 1:4)
}
