% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.ds.R
\name{pattern.ds}
\alias{pattern.ds}
\title{Pattern : Responses in descending order}
\usage{
pattern.ds(data, cols)
}
\arguments{
\item{data}{The data set with the scored responses in the same order as it is administered}

\item{cols}{The columns belonging to the psychometric scale/scales under scrutiny}
}
\value{
The participants (rows) following the pattern, and
the list of said participants along with their responses for all the items in the scale
}
\description{
Pattern : Responses in descending order
}
\examples{
pattern.ds(data, 13:24)
}
